Tidied up a bunch of files

+++ ns-3-dev/src/processing/model/membus.cc
@@ -30,8 +30,7 @@ void
 MemBus::Contend(ProcessingInstance *pi)
 {
   m_contenders.push_back(pi);
-  pi->remaining[MEMSTALLCYCLES].amount = 
-    freq / m_contenders.size() * (pi->remaining[MEMORYACCESSES].amount);
+  pi->remaining[MEMSTALLCYCLES].amount = (float)freq / m_contenders.size() * (pi->remaining[MEMORYACCESSES].amount);
 }
 
 // NEXT: MemBus::UnContend
